+++ lib/config.py
@@ -68,6 +68,7 @@ CONFIG_SCHEMA = {
         '__type__': 'dict',
         'use_numstat':         {'type': 'bool'},
         'count_hunks':         {'type': 'bool'},
+        'cherry_pick_test':    {'type': 'bool'},
         'no_merges':           {'type': 'bool'},
         'first_parent':        {'type': 'bool'},
         'score_workers':       {'type': 'int'},